Refactored and added test to test old payment service. Will use this to verify new payment service does it's job.
