+++ payment/payment.go
@@ -1,9 +1,10 @@
-package main
+package payment
 
 import (
 	"flag"
 	"fmt"
 	"net/http"
+	"io"
 )
 
 var port string
@@ -13,12 +14,13 @@ func main() {
 	flag.StringVar(&port, "port", "8082", "Port on which to run")
 	flag.Parse()
 
-	http.HandleFunc("/paymentAuth", paymentAuthHandler)
+	http.HandleFunc("/paymentAuth", PaymentAuthHandler)
 	fmt.Printf("Payment service running on port %s\n", port)
 	http.ListenAndServe(":"+port, nil)
 }
 
-func paymentAuthHandler(w http.ResponseWriter, r *http.Request) {
-	fmt.Printf("Received payment auth request... Authorized.\n")
-	w.WriteHeader(200)
+func PaymentAuthHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+	w.Header().Set("Content-Type", "application/json")
+	io.WriteString(w, `{"authorised": true}`)
 }
